Poll pages with redirect to questions

+++ src/main/java/com/eucolus/poll/repositories/RequestBrowserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 
-public interface RequestBrowserRepository extends CrudRepository<RequestBrowser, Long> {
+public interface RequestBrowserRepository extends CrudRepository<RequestBrowser, Integer> {
     @Query(value = "SELECT * FROM request_browsers WHERE name=(:name) AND version=(:version)", nativeQuery = true)
     RequestBrowser find(@Param("name") String name, @Param("version") String version);
 }
